requirements/requirements.base.txt
pillow
numpy
tqdm
scikit-learn
matplotlib
pillow
pandas